ItemAttributes: Improved recognition; Refactors + moved stuff to proper folders; Decreased log verbosity; Made player spawn naked; Added random uniprefabs to TestMap
