+++ Assets/Prefabs/Player/Resources/Player_V3(uNet).prefab
@@ -1353,20 +1353,19 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   suitPrefab: {fileID: 0}
-  beltPrefab: {fileID: 1816790622172278, guid: 74303457ff5f1a9449b6c406f1ac2801, type: 2}
+  beltPrefab: {fileID: 0}
   shoesPrefab: {fileID: 0}
   headPrefab: {fileID: 0}
   maskPrefab: {fileID: 0}
-  uniformPrefab: {fileID: 1011433845357754, guid: ff69ce97a90a44bb58258ee9d8fad584,
-    type: 2}
+  uniformPrefab: {fileID: 0}
   neckPrefab: {fileID: 0}
-  earPrefab: {fileID: 1469499218258322, guid: 2eae96f88d2a8d64987a4867bd21fbbf, type: 2}
+  earPrefab: {fileID: 0}
   glassesPrefab: {fileID: 0}
   glovesPrefab: {fileID: 0}
   leftHandPrefab: {fileID: 0}
   rightHandPrefab: {fileID: 0}
-  idPrefab: {fileID: 1748120534729806, guid: 4c73033e5f8ab734eb2e85e6ae54de73, type: 2}
-  bagPrefab: {fileID: 1426788842454498, guid: 7a7587eaf2b5e1c44addf7d2475d86c9, type: 2}
+  idPrefab: {fileID: 0}
+  bagPrefab: {fileID: 0}
   storage01Prefab: {fileID: 0}
   storage02Prefab: {fileID: 0}
   suitStoragePrefab: {fileID: 0}
