+++ Assets/Resources/DmiIconData.asset.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 62b22ae789cae4c68b62c8b908b1137a
-timeCreated: 1496776361
+guid: ad8f7ae17cd514a22a18d4ffc482c8c8
+timeCreated: 1496953722
 licenseType: Pro
 NativeFormatImporter:
   mainObjectFileID: 11400000
