+++ Assets/scripts/Data/DmiIconData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 7fd860f3a4f144ba2bbfa92aa2a33d3f
+timeCreated: 1496388099
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
