+++ Assets/scripts/Data/DmiState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c0de5cfda00041b2bedcbcebb8e89b5d
+timeCreated: 1496495994
