+++ Assets/scripts/EditorTools/SpriteSlicer.cs
