--- Assets/scripts/Test.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: d83459c9f5a6840349009a09b0e5e895
-folderAsset: yes
-timeCreated: 1496299133
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
