--- Assets/scripts/Test/DmiIcon.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 91427411514a484190ba8275e4686716
-timeCreated: 1496495978
