--- Assets/scripts/Test/DmiState.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: c0de5cfda00041b2bedcbcebb8e89b5d
-timeCreated: 1496495994
