--- Assets/scripts/Test/SpriteSlicer.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d504559c7a40f4a6f9829e9ef2e1d7ce
-timeCreated: 1496398700
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
