--- Assets/scripts/Test/TgObjectTest.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 397d52b9e50e418a928ad73f012c51fe
-timeCreated: 1496314548
