--- Assets/scripts/Test/TgSpriteTest.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d3f3043b021f9446b8153a3991d4a7fb
-timeCreated: 1496299190
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
